Moved Prime_numGen Project into sub folder.
